requirements.txt
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.1
Django==2.1.3
django-cors-headers==2.4.0
django-oidc-provider==0.7.0
django-user-sessions==1.6.0
djangorestframework==3.9.0
ecdsa==0.13
future==0.17.1
idna==2.7
jwcrypto==0.6.0
more-itertools==4.3.0
names==0.3.0
pkg-resources==0.0.0
pluggy==0.8.0
py==1.7.0
pycparser==2.19
pycryptodome==3.7.2
pycryptodomex==3.7.0
pyjwkest==1.4.0
PyJWT==1.6.4
pytest==4.0.1
pytz==2018.7
PyYAML==3.13
requests==2.20.1
six==1.11.0
urllib3==1.24.1
uWSGI==2.0.17.1
